Fix keys being optional on writes
