+++ .changeset/pink-wolves-sleep.md
@@ -1,0 +1,5 @@
+---
+'@directus/api': patch
+---
+
+Fix keys being optional on writes
